move Integer/Long.getChars to StringLatin1
